[Refactor] 코테 커스텀 훅 리팩토링
