+++ src/pages/coding-test/CodingTest.tsx
@@ -3,7 +3,7 @@ import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
 import EditorPanel from '@/pages/coding-test/components/EditorPanel';
 import ProblemPanel from '@/pages/coding-test/components/ProblemPanel';
 import Header from '@/pages/coding-test/components/Header';
-import { useCodingTestLogic } from '@/pages/coding-test/hooks/useCodingTestLogic';
+import { useCodingTestMainLogic } from '@/pages/coding-test/hooks/useCodingTestMainLogic';
 import { EditorTheme } from '@/pages/coding-test/types/types';
 import '@/css/coding-test/CodingTest.scss';
 
@@ -19,7 +19,7 @@ const CodingTest= () => {
     handleRunCode,
     handleSubmit,
     currentCode
-  } = useCodingTestLogic();
+  } = useCodingTestMainLogic();
 
   return (
     <div className="coding-test-container dark-mode">
